Fix data type conversion to int32 for model training

+++ VishwamAI/scripts/train_vishwamai_model.py
@@ -70,14 +70,14 @@ def train_model(data_file, num_epochs=10):
     # Initialize model parameters
     example_batch = next(iter(dataset))
     example_batch = example_batch.numpy().tolist()  # Convert tensor to list of lists of integers
-    example_batch = jax.numpy.array(example_batch, dtype=jnp.float32)  # Convert to float32
+    example_batch = jax.numpy.array(example_batch, dtype=jnp.int32)  # Convert to int32
     params = model.init(rng, example_batch)
 
     # Training loop
     for epoch in range(num_epochs):
         for batch in dataset:
             batch = batch.numpy().tolist()  # Convert tensor to list of lists of integers
-            batch = jax.numpy.array(batch, dtype=jnp.float32)  # Convert to float32
+            batch = jax.numpy.array(batch, dtype=jnp.int32)  # Convert to int32
             loss, params, opt_state = train_step(params, model, optimizer, batch, rng)
             logging.info(f"Epoch {epoch + 1}, Loss: {loss}")
